update main.py, update readme for input arguments
