+++ main.py
@@ -2,13 +2,18 @@ import os
 import argparse
 import sympy as sp
 from symbolic_tensor_graph.graph.graph import TensorGraph
-from symbolic_tensor_graph.graph.grad_updater import GradUpdater
+from symbolic_tensor_graph.graph.grad_updater import (
+    GradUpdater,
+    MicroBatchReplicator,
+    MicroBatchReplicatorPostProcess,
+)
+from symbolic_tensor_graph.graph.replicate_graph import ReplicateGraph
 from symbolic_tensor_graph.graph.graph_distributer import GraphDistributer
 from symbolic_tensor_graph.graph.convert_chakra import BundledConvertChakra
-from models.transformer import (
-    transformer_stack as transformer_stack_fn, 
-    transformer as transformer_fn
-)
+import re
+from symbolic_tensor_graph.vram_counting import _print_gpu_vram
+
+mixprecision = False
 
 
 def str_to_bool(v):
@@ -16,143 +21,548 @@ def str_to_bool(v):
     return v.lower() in ("true", "t", "1", "yes", "y")
 
 
+def _create_pipeline_tensor_map_mix_precision(
+    _tensors, _temporal_parallel_dims, _symbol_map_value, num_stacks
+):
+    _tensor_map = dict()
+    assert len(_temporal_parallel_dims) == 1
+    parallel_dim = _temporal_parallel_dims[0]
+    range_ = _symbol_map_value[parallel_dim]
+
+    # Determine how many transformer blocks belong to each pipeline stage
+    num_stacks_each_stage = [num_stacks // range_] * range_
+    for i in range(num_stacks % range_):
+        num_stacks_each_stage[i] += 1  # distribute remainder to early stages
+    # Cumulative upper bounds for easy stage lookup
+    cumulative = []
+    acc = 0
+    for v in num_stacks_each_stage:
+        acc += v
+        cumulative.append(acc)
+
+    for tensor in _tensors:
+        tid = tensor.id
+        # ------------------------------------------------------------------
+        # 1) Transformer block tensors
+        # ------------------------------------------------------------------
+        m = re.search(r"transformer\.(\d+)", tid)
+        if m:
+            block_idx = int(m.group(1))
+            # Find the first cumulative upper bound that exceeds block_idx
+            stage = next(i for i, up in enumerate(cumulative) if block_idx < up)
+            _tensor_map[tid] = {parallel_dim: stage}
+            continue
+
+        # ------------------------------------------------------------------
+        # 2) Special tensors (embeddings, loss etc.)
+        # ------------------------------------------------------------------
+        if "in_emb" in tid:
+            _tensor_map[tid] = {parallel_dim: 0}
+        elif "out_emb" in tid or "loss" in tid:
+            _tensor_map[tid] = {parallel_dim: (range_ - 1)}
+        else:
+            # Any tensor that doesn't match the above categories should be
+            # impossible – raise explicit error to catch new patterns early.
+            raise ValueError(f"Unrecognized tensor id for pipeline mapping: {tid}")
+
+    return _tensor_map
+
+
+def _create_pipeline_tensor_map(
+    _tensors, _temporal_parallel_dims, _symbol_map_value, num_stacks
+):
+    if mixprecision:
+        return _create_pipeline_tensor_map_mix_precision(
+            _tensors, _temporal_parallel_dims, _symbol_map_value, num_stacks
+        )
+    _tensor_map = dict()
+    assert len(_temporal_parallel_dims) == 1
+    parallel_dim = _temporal_parallel_dims[0]
+    range_ = _symbol_map_value[parallel_dim]
+    num_stacks_each_stage = list()
+    for i in range(range_):
+        num_stacks_each_stage.append(num_stacks // range_)
+    for i in range(num_stacks - range_ * (num_stacks // range_)):
+        num_stacks_each_stage[i] += 1
+    for i in range(range_):
+        if i == 0:
+            continue
+        num_stacks_each_stage[i] += num_stacks_each_stage[i - 1]
+    # num_stacks_each_stage.append(num_stacks_each_stage[-1]+100000)
+
+    for tensor in _tensors:
+        if tensor.id == "transformer.18._sharded_weight@1":
+            pass
+        found = False
+        for num_stack in range(num_stacks):
+            if f"transformer.{num_stack}." in tensor.id:
+                for stage, upper_bound in enumerate(num_stacks_each_stage):
+                    if num_stack < upper_bound:
+                        _tensor_map[tensor.id] = {parallel_dim: stage}
+                        found = True
+                        break
+                if found:
+                    break
+        if found:
+            pass
+        elif "in_emb" in tensor.id:
+            _tensor_map[tensor.id] = {parallel_dim: 0}
+        elif "out_emb" in tensor.id:
+            _tensor_map[tensor.id] = {parallel_dim: (num_stacks - 1) % range_}
+        elif "loss" in tensor.id:
+            _tensor_map[tensor.id] = {parallel_dim: (num_stacks - 1) % range_}
+        else:
+            assert False, tensor.name
+    return _tensor_map
+
+
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument("--output_dir", type=str, help="dir where stores output traces", required=True)
-    parser.add_argument("--output_name", type=str, help="name of output traces", required=True)
-    parser.add_argument("--comm_group_file", type=str, help="name of comm_group_file", required=True)
-    parser.add_argument("--dp", type=int, help="data parallel degree", required=False, default=1)
-    parser.add_argument("--tp", type=int, help="tensor parallel degree", required=False, default=1)
-    parser.add_argument("--sp", type=int, help="sequence parallel degree", required=False, default=1)
-    parser.add_argument("--pp", type=int, default=1, help="pipeline parallel degree", required=False)
-    parser.add_argument("--weight_sharded", type=str_to_bool, help="whether weight sharded", required=False, default=False)
-    parser.add_argument("--activation_recompute", type=str_to_bool, help="whether recompute activation", required=False, default=False)
-    parser.add_argument("--din", type=int, default=51200, required=False)
-    parser.add_argument("--dout", type=int, default=25600, required=False)
-    parser.add_argument("--dmodel", type=int, default=25600, required=False)
-    parser.add_argument("--dff", type=int, default=25600*4, required=False)
-    parser.add_argument("--batch", type=int, default=1024, required=False)
+    parser.add_argument(
+        "--output_dir", type=str, help="dir where stores output traces", required=True
+    )
+    parser.add_argument(
+        "--output_name", type=str, help="name of output traces", required=True
+    )
+    parser.add_argument(
+        "--dp", type=int, help="data parallel degree", required=False, default=1
+    )
+    parser.add_argument(
+        "--tp", type=int, help="tensor parallel degree", required=False, default=1
+    )
+    parser.add_argument(
+        "--sp", type=int, help="sequence parallel degree", required=False, default=1
+    )
+    parser.add_argument(
+        "--ep", type=int, help="expert parallel degree", required=False, default=1
+    )
+    parser.add_argument(
+        "--pp", type=int, default=1, help="pipeline parallel degree", required=False
+    )
+    parser.add_argument(
+        "--weight_sharded",
+        type=str_to_bool,
+        help="whether weight sharded",
+        required=False,
+        default=False,
+    )
+    parser.add_argument(
+        "--activation_recompute",
+        type=str_to_bool,
+        help="whether recompute activation",
+        required=False,
+        default=False,
+    )
+    parser.add_argument(
+        "--tpsp",
+        type=str_to_bool,
+        help="use tp+sp or tp only",
+        required=False,
+        default=True,
+    )
+    parser.add_argument("--dvocal", type=int, default=32000, required=False)
+    parser.add_argument("--dmodel", type=int, default=8192, required=False)
+    parser.add_argument("--dff", type=int, default=28672, required=False)
+    parser.add_argument("--batch", type=int, default=64, required=False)
+    parser.add_argument("--micro_batch", type=int, default=-1, required=False)
     parser.add_argument("--seq", type=int, default=1024, required=False)
-    parser.add_argument("--head", type=int, default=1024, required=False)
-    parser.add_argument("--num_stacks", type=int, default=32, required=False)
-    parser.add_argument("--chakra_schema_version", type=str, default="v0.0.4", required=False)
-    parser.add_argument("--mixed_precision", type=str_to_bool, default=False, required=False)
-    
+    parser.add_argument("--head", type=int, default=64, required=False)
+    parser.add_argument("--kvhead", type=int, default=8, required=False)
+    parser.add_argument("--num_stacks", type=int, default=80, required=False)
+    parser.add_argument("--experts", type=int, default=8, required=False)
+    parser.add_argument("--kexperts", type=int, default=2, required=False)
+    parser.add_argument(
+        "--chakra_schema_version", type=str, default="v0.0.4", required=False
+    )
+    parser.add_argument("--model_type", type=str, default="dense", required=False)
+    parser.add_argument(
+        "--mixed_precision", type=str_to_bool, default=False, required=False
+    )
+    parser.add_argument(
+        "--print_gpu_vram",
+        type=str_to_bool,
+        default=False,
+        required=False,
+        help="Whether to print per-GPU VRAM footprint (total / params / acts / grads) in GiB",
+    )
+
     args = parser.parse_args()
 
     os.makedirs(args.output_dir, exist_ok=True)
     if not "%d" in args.output_name:
         args.output_name = f"{args.output_name}.%d.et"
     generated_filename = os.path.join(args.output_dir, args.output_name)
-    dp, tp, pp, spp = sp.symbols("dp tp pp sp")
-    Din, Dout, Dmodel, Dff, Batch, Seq, Head = sp.symbols(
-        "Din Dout Dmodel Dff Batch Seq Head"
+    dp, tp, pp, spp, ep, fsdp = sp.symbols("dp tp pp cp ep fsdp")
+    (
+        Din,
+        Dout,
+        Dmodel,
+        Dff,
+        Batch,
+        Seq,
+        Head,
+        KVHead,
+        Experts,
+        KExperts,
+        Dvocal,
+        MicroBatch,
+    ) = sp.symbols(
+        "Din Dout Dmodel Dff Batch Seq Head KVHead Experts KExperts Dvocal MicroBatch"
     )
+    if args.micro_batch == -1:
+        args.micro_batch = args.batch
     symbol_map_value = {
-        Din: args.din,
-        Dout: args.dout,
+        Dvocal: args.dvocal,
         Dmodel: args.dmodel,
         Dff: args.dff,
         Batch: args.batch,
+        MicroBatch: args.micro_batch,
         Seq: args.seq,
         Head: args.head,
+        KVHead: args.kvhead,
+        Experts: args.experts,
+        KExperts: args.kexperts,
         dp: args.dp,
         tp: args.tp,
         pp: args.pp,
-        spp: args.sp
+        spp: args.sp,
+        ep: args.ep,
     }
     num_stacks = args.num_stacks
-    spatial_parallel_dims = [dp, tp, spp]
     temporal_parallel_dims = [pp]
-    
-    if not args.activation_recompute:
-        if not args.weight_sharded:
-            module_template_dir = os.path.join(
-                    os.path.split(
-                        os.path.abspath(__file__)
-                    )[0],
-                    "./sharding_spreadsheets/module/fullset"  
+    if args.weight_sharded:
+        symbol_map_value[fsdp] = args.dp if args.dp != 0 else 1
+        symbol_map_value["fsdp"] = args.dp if args.dp != 0 else 1
+    else:
+        symbol_map_value[fsdp] = 1
+        symbol_map_value["fsdp"] = 1
+
+    hook = 1
+    global mixprecision
+    if args.mixed_precision:
+        mixprecision = True
+
+    if args.model_type == "llama" or args.model_type == "dense":
+        if mixprecision:
+            from models.stage1.llama_model import llama as transformer_dense
+        else:
+            from models.stage1.gpt_model import gpt as transformer_dense
+
+        print("Assembling dense model")
+        transformer_dense = transformer_dense(
+            num_stacks, regenerate=True, tpsp=args.tpsp
+        )
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") == "0":
+            transformer_dense = MicroBatchReplicator.apply(
+                transformer_dense, symbol_map_value
             )
         else:
-            module_template_dir = os.path.join(
-                    os.path.split(
-                        os.path.abspath(__file__)
-                    )[0],
-                    "./sharding_spreadsheets/module/fully_sharded_fullset"
+            print("[Warning] MICROBATCH OPTIMIZE sometimes generate incorrect graphs, use with caution!")
+            transformer_dense = ReplicateGraph.apply(
+                transformer_dense,
+                inplace=True,
+                old_symbol_map_new_symbol={"Batch": "MicroBatch"},
             )
-    else:
-        if not args.weight_sharded:
-            module_template_dir = os.path.join(
-                    os.path.split(
-                        os.path.abspath(__file__)
-                    )[0],
-                    "./sharding_spreadsheets/module/fullset_recomp"  
+
+        if args.weight_sharded:
+            transformer_dense = ReplicateGraph.apply(
+                transformer_dense,
+                inplace=True,
+                old_symbol_map_new_symbol={"fsdp": "dp"},
             )
         else:
-            module_template_dir = os.path.join(
-                    os.path.split(
-                        os.path.abspath(__file__)
-                    )[0],
-                    "./sharding_spreadsheets/module/fully_sharded_fullset_recomp"
+            transformer_dense = ReplicateGraph.apply(
+                transformer_dense, inplace=True, old_symbol_map_new_symbol={"fsdp": 1}
             )
-        
-    # build the tensor graph
-    mha = TensorGraph.load_tensor_graph(
-        os.path.join(module_template_dir, "multi_head_attention.csv")
-    )
-    ffn = TensorGraph.load_tensor_graph(
-        os.path.join(module_template_dir, "feed_forward_network.csv")
-    )
-    in_emb = TensorGraph.load_tensor_graph(
-        os.path.join(module_template_dir, "embedding.csv")
-    )
-    out_emb = TensorGraph.load_tensor_graph(
-        os.path.join(module_template_dir, "embedding.csv")
-    )
-    stack = transformer_stack_fn(mha, ffn)
-    transformer = transformer_fn(in_emb, out_emb, stack, num_stacks)
-    transformer_updated_grad = GradUpdater.apply(transformer)
-    
-    # transformer.visualize(os.path.join(args.output_dir, "transformer"), format="pdf")
-    
-    # distribute tensor graph to machines
-    def _create_pipeline_tensor_map(_tensors, _temporal_parallel_dims, _symbol_map_value):
-        _tensor_map = dict()
-        assert len(_temporal_parallel_dims) == 1
-        parallel_dim = _temporal_parallel_dims[0]
-        range_ = _symbol_map_value[parallel_dim]
-        for tensor in _tensors:
-            for num_stack in range(num_stacks):
-                if f"stack_{num_stack}_" in tensor.id:
-                    _tensor_map[tensor.id] = {parallel_dim: (num_stack+1) % range_}
-                    break
-            if "in_emb" in tensor.id:
-                _tensor_map[tensor.id] = {parallel_dim: 0}
-            elif "out_emb" in tensor.id:
-                _tensor_map[tensor.id] = {parallel_dim: (num_stacks+1) % range_}
-        return _tensor_map
-    pipeline_tensor_map = _create_pipeline_tensor_map(transformer_updated_grad.tensors, temporal_parallel_dims, symbol_map_value)
-    distributed_tensor_graph = GraphDistributer.apply(
-        transformer_updated_grad,
-        symbol_map_value,
-        spatial_parallel_dims,
-        temporal_parallel_dims,
-        pipeline_tensor_map
-    )
-    
-    # readout to chakra
-    distributed_chakra_graph = BundledConvertChakra.apply(distributed_tensor_graph, symbol_map_value, os.path.join(args.output_dir, args.comm_group_file), mixed_precision=args.mixed_precision)
-    if args.chakra_schema_version == "v0.0.1":
-        from symbolic_tensor_graph.chakra.backends.chakra_00_1_backend import Chakra001Backend as ReadoutBackend
-    elif args.chakra_schema_version == "v0.0.4":
-        from symbolic_tensor_graph.chakra.backends.chakra_00_4_backend import Chakra004Backend as ReadoutBackend
-    elif args.chakra_schema_version == "json":
-        from symbolic_tensor_graph.chakra.backends.json_backend import JsonBackend as ReadoutBackend
-    else:
-        assert False
-    distributed_chakra_graph.readout(generated_filename, backend=ReadoutBackend)
+
+        # transformer_dense.visualize("dense")
+        # transformer_dense.save_tensor_graph("llama.csv")
+
+        transformer_dense = GradUpdater.apply(transformer_dense, inplace=True)
+        spatial_parallel_dims_dense = [dp, tp, spp]
+
+        symbol_map_value[tp] *= symbol_map_value[ep]
+        # dense model
+        pipeline_tensor_map = _create_pipeline_tensor_map(
+            transformer_dense.tensors,
+            temporal_parallel_dims,
+            symbol_map_value,
+            num_stacks,
+        )
+
+        print("Dense model: Distributing")
+        distributed_tensor_graph_dense = GraphDistributer.apply(
+            transformer_dense,
+            symbol_map_value,
+            spatial_parallel_dims_dense,
+            temporal_parallel_dims,
+            pipeline_tensor_map,
+        )
+
+        if args.print_gpu_vram:
+            _print_gpu_vram(
+                distributed_tensor_graph_dense,
+                symbol_map_value,
+                mixed_precision=args.mixed_precision,
+                header="[Dense] ",
+            )
+
+        print("Dense model: Converting Chakra")
+        comm_group_file = args.output_name.replace(".%d", "").replace(".et", ".json")
+        distributed_chakra_graph_dense = BundledConvertChakra.apply(
+            distributed_tensor_graph_dense,
+            symbol_map_value,
+            os.path.join(args.output_dir, comm_group_file),
+            mixed_precision=args.mixed_precision,
+        )
+
+        from symbolic_tensor_graph.chakra.backends.chakra_00_4_backend import (
+            Chakra004Backend as ReadoutBackend,
+        )
+
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") != "0":
+            distributed_chakra_graph_dense = MicroBatchReplicatorPostProcess.apply(
+                distributed_chakra_graph_dense, args.batch // args.micro_batch
+            )
+
+        print("Dense model: reading out")
+        distributed_chakra_graph_dense.readout(
+            generated_filename, backend=ReadoutBackend
+        )
+    elif args.model_type == "gpt":
+        from models.stage1.gpt_model import gpt as transformer_dense
+
+        print("Assembling dense model")
+        transformer_dense = transformer_dense(
+            num_stacks, regenerate=True, tpsp=args.tpsp
+        )
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") == "0":
+            transformer_dense = MicroBatchReplicator.apply(
+                transformer_dense, symbol_map_value
+            )
+        else:
+            print("[Warning] MICROBATCH OPTIMIZE sometimes generate incorrect graphs, use with caution!")
+            transformer_dense = ReplicateGraph.apply(
+                transformer_dense,
+                inplace=True,
+                old_symbol_map_new_symbol={"Batch": "MicroBatch"},
+            )
+
+        if args.weight_sharded:
+            transformer_dense = ReplicateGraph.apply(
+                transformer_dense,
+                inplace=True,
+                old_symbol_map_new_symbol={"fsdp": "dp"},
+            )
+        else:
+            transformer_dense = ReplicateGraph.apply(
+                transformer_dense, inplace=True, old_symbol_map_new_symbol={"fsdp": 1}
+            )
+
+        # transformer_dense.visualize("dense")
+        # transformer_dense.save_tensor_graph("gpt.csv")
+
+        transformer_dense = GradUpdater.apply(transformer_dense, inplace=True)
+        spatial_parallel_dims_dense = [dp, tp, spp]
+
+        symbol_map_value[tp] *= symbol_map_value[ep]
+        # dense model
+        pipeline_tensor_map = _create_pipeline_tensor_map(
+            transformer_dense.tensors,
+            temporal_parallel_dims,
+            symbol_map_value,
+            num_stacks,
+        )
+
+        print("Dense model: Distributing")
+        distributed_tensor_graph_dense = GraphDistributer.apply(
+            transformer_dense,
+            symbol_map_value,
+            spatial_parallel_dims_dense,
+            temporal_parallel_dims,
+            pipeline_tensor_map,
+        )
+
+        if args.print_gpu_vram:
+            _print_gpu_vram(
+                distributed_tensor_graph_dense,
+                symbol_map_value,
+                mixed_precision=args.mixed_precision,
+                header="[GPT] ",
+            )
+
+        print("Dense model: Converting Chakra")
+        comm_group_file = args.output_name.replace(".%d", "").replace(".et", ".json")
+        distributed_chakra_graph_dense = BundledConvertChakra.apply(
+            distributed_tensor_graph_dense,
+            symbol_map_value,
+            os.path.join(args.output_dir, comm_group_file),
+            mixed_precision=args.mixed_precision,
+        )
+
+        from symbolic_tensor_graph.chakra.backends.chakra_00_4_backend import (
+            Chakra004Backend as ReadoutBackend,
+        )
+
+        print("Dense model: reading out")
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") != "0":
+            distributed_chakra_graph_dense = MicroBatchReplicatorPostProcess.apply(
+                distributed_chakra_graph_dense, args.batch // args.micro_batch
+            )
+        distributed_chakra_graph_dense.readout(
+            generated_filename, backend=ReadoutBackend
+        )
+
+    elif args.model_type == "moe":
+        from models.stage1.moe_model import transformer as transformer_moe
+
+        assert args.tpsp
+        print("Assembling moe model")
+        transformer_moe = transformer_moe(num_stacks, symbol_map_value, regenerate=True)
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") == "0":
+            transformer_moe = MicroBatchReplicator.apply(
+                transformer_moe, symbol_map_value
+            )
+        else:
+            print("[Warning] MICROBATCH OPTIMIZE sometimes generate incorrect graphs, use with caution!")
+            assert False, "disable for now"
+        transformer_moe = ReplicateGraph.apply(
+            transformer_moe,
+            inplace=True,
+            old_symbol_map_new_symbol={"Batch": "MicroBatch"},
+        )
+
+        if args.weight_sharded:
+            transformer_moe = ReplicateGraph.apply(
+                transformer_moe,
+                inplace=True,
+                old_symbol_map_new_symbol={"fsdp": "dp"},
+            )
+        else:
+            transformer_moe = ReplicateGraph.apply(
+                transformer_moe, inplace=True, old_symbol_map_new_symbol={"fsdp": 1}
+            )
+
+        # transformer_moe.visualize("moe")
+        # transformer_moe.save_tensor_graph("moe.csv")
+        transformer_moe = GradUpdater.apply(transformer_moe, inplace=True)
+        spatial_parallel_dims_moe = [dp, tp, spp, ep]
+
+        # moe model
+        pipeline_tensor_map = _create_pipeline_tensor_map(
+            transformer_moe.tensors,
+            temporal_parallel_dims,
+            symbol_map_value,
+            num_stacks,
+        )
+
+        print("MoE model: Distributing")
+        distributed_tensor_graph_moe = GraphDistributer.apply(
+            transformer_moe,
+            symbol_map_value,
+            spatial_parallel_dims_moe,
+            temporal_parallel_dims,
+            pipeline_tensor_map,
+        )
+
+        if args.print_gpu_vram:
+            _print_gpu_vram(
+                distributed_tensor_graph_moe,
+                symbol_map_value,
+                mixed_precision=args.mixed_precision,
+                header="[MoE] ",
+            )
+
+        print("MoE model: Converting Chakra")
+        comm_group_file = args.output_name.replace(".%d", "").replace(".et", ".json")
+        distributed_chakra_graph_moe = BundledConvertChakra.apply(
+            distributed_tensor_graph_moe,
+            symbol_map_value,
+            os.path.join(args.output_dir, comm_group_file),
+            mixed_precision=args.mixed_precision,
+        )
+
+        from symbolic_tensor_graph.chakra.backends.chakra_00_4_backend import (
+            Chakra004Backend as ReadoutBackend,
+        )
+
+        print("MoE model: reading out")
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") != "0":
+            distributed_chakra_graph_moe = MicroBatchReplicatorPostProcess.apply(
+                distributed_chakra_graph_moe, args.batch // args.micro_batch
+            )
+        distributed_chakra_graph_moe.readout(generated_filename, backend=ReadoutBackend)
+
+    elif args.model_type == "debug":
+        transformer_moe = TensorGraph.load_tensor_graph(
+            "./sharding_spreadsheets/module3/tpsp/embedding.csv"
+        )
+        transformer_moe = ReplicateGraph.apply(
+            transformer_moe,
+            inplace=True,
+            old_symbol_map_new_symbol={
+                "Batch": "MicroBatch",
+                "Din": "Dvocal",
+                "Dout": "Dvocal",
+            },
+        )
+
+        if args.weight_sharded:
+            transformer_moe = ReplicateGraph.apply(
+                transformer_moe,
+                inplace=True,
+                old_symbol_map_new_symbol={"fsdp": "dp"},
+            )
+        else:
+            transformer_moe = ReplicateGraph.apply(
+                transformer_moe, inplace=True, old_symbol_map_new_symbol={"fsdp": 1}
+            )
+
+        # transformer_moe.visualize("moe")
+        # transformer_moe.save_tensor_graph("moe.csv")
+        transformer_moe = GradUpdater.apply(transformer_moe, inplace=True)
+        spatial_parallel_dims_moe = [dp, tp, spp, ep]
+
+        # moe model
+        assert args.pp == 1
+        pipeline_tensor_map = {
+            "x@0": {pp: 0},
+            "w@0": {pp: 0},
+            "y@0": {pp: 0},
+            "dy@0": {pp: 0},
+            "dw@0": {pp: 0},
+            "dx@0": {pp: 0},
+            "w@1": {pp: 0},
+        }
+
+        print("MoE model: Distributing")
+        distributed_tensor_graph_moe = GraphDistributer.apply(
+            transformer_moe,
+            symbol_map_value,
+            spatial_parallel_dims_moe,
+            temporal_parallel_dims,
+            pipeline_tensor_map,
+        )
+
+        print("MoE model: Converting Chakra")
+        comm_group_file = args.output_name.replace(".%d", "").replace(".et", ".json")
+        distributed_chakra_graph_moe = BundledConvertChakra.apply(
+            distributed_tensor_graph_moe,
+            symbol_map_value,
+            os.path.join(args.output_dir, comm_group_file),
+        )
+
+        from symbolic_tensor_graph.chakra.backends.chakra_00_4_backend import (
+            Chakra004Backend as ReadoutBackend,
+        )
+
+        print("MoE model: reading out")
+        if os.environ.get("STAGE_MICROBATCH_OPTIMIZE", "0") != "0":
+            distributed_chakra_graph_moe = MicroBatchReplicatorPostProcess.apply(
+                distributed_chakra_graph_moe, args.batch // args.micro_batch
+            )
+        distributed_chakra_graph_moe.readout(generated_filename, backend=ReadoutBackend)
 
 
 if __name__ == "__main__":
     main()
-
